+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 
 
 ## 2026-04-18
+- `23:05` -- **Updated** `apps/desktopProbe/src/server/overlayBrowserView.ts`
 - `23:04` -- **Updated** `apps/desktopProbe/src/server/jobScanner.ts`
 - `23:03` -- **Updated** `apps/desktopProbe/src/server/jobScanner.ts`
 - `22:04` -- **Updated** `apps/desktopProbe/src/server/overlayBrowserView.ts`